Sketch out login
The basic system should work now. Data still needs to be fetched etc.
but the hard part (auth) seems to be working.

It looks like it is enough simply to use InAppBrowser for OAuth. The
parent browser will then share its auth state and subsequent requests
will be authed as they should.